+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- ajaxOptions option is now passed the MultiListControl instance
+
 ## [0.3.0] - 2019-10-28
 
 ### Added